Rename FrozenResource -> Resource, Resource -> MutableResource
These names are a bit more intuitive
